development/server/requirements.txt
bcolz==1.1.2
boto==2.48.0
bz2file==0.98
certifi==2017.4.17
chardet==3.0.4
decorator==4.0.11
gensim==2.2.0
h5py==2.7.0
idna==2.5
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
Keras==2.0.6
mock==2.0.0
nltk==3.2.4
numpy==1.13.1
olefile==0.44
pandas==0.20.2
pbr==3.1.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.2.1
prompt-toolkit==1.0.14
protobuf==3.3.0
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.6.0
pytz==2017.2
PyYAML==3.12
requests==2.18.1
scikit-learn==0.18.2
scipy==0.19.1
simplegeneric==0.8.1
six==1.10.0
smart-open==1.5.3
tensorflow==1.3.0
traitlets==4.3.2
urllib3==1.21.1
wcwidth==0.1.7
Werkzeug==0.12.2
xgboost==0.6a2
aniso8601==1.2.1
click==6.7
Flask==0.12.2
flask-restplus==0.10.1
fluent-logger==0.5.0
itsdangerous==0.24
Jinja2==2.9.6
jsonschema==2.6.0
MarkupSafe==1.0
msgpack-python==0.4.8
python-dateutil==2.6.0
pytz==2017.2
simplejson==3.10.0
Werkzeug==0.12.2